Match belt/processor etc. level to upgrade tiers
Currently the belt (+ processor) levels do not match to the upgrade shop tiers. The tiers start from Tier 1, and once upgraded, they go to tier 2, 3, ... 7. So tier 7 is currently the highest.

However, the belt level starts from level 0, and increases to level 1 once upgraded. The maximum level is therefore 6.

This PR raises the start level of belts and other items to level 1, meaning the levels match the upgrade tiers in the shop.