Update base-de.yaml
Translation of the new content and improvement to old/already existing translation.

Also rearranged the translation keys so it matches english version and trimmed all free rows (Is this needed? I know the translation file got very long, but now it's kind of one mishmash and hard to read).

Some of the rewards are still missing, but I will complete it before October 8th.