Update base-de.yaml (#786)
This should complete the german translation for the new update. I took some inspiration from @everthingArduino's propositions. As always, it would be nice if someone has an addidional pair of eyes on this 👀 Just in case. Maybe someone can have a look at the german unlock message of virtual processing who knows how to build from GitHub? I have a feeling it might be too long 😅 
However, I'm excited for the new update and I wish a good and smooth launch to @tobspr this friday! 🥳