Allow for compasses for custom waypoints
Waypoints that begin with "!" (or another prefix specified by COMPASS_PREFIX)
will display a compass that functions identically to the hub compass.

Because compasses are remembered by position, two markers in the same position
will share a compass, leading to other small issues.

Camera recentering is also now exact, with a slight change to how it operates.