Increase minimum screenshot scale
The pixels-per-tile for screenshots must now be an odd integer of at least 3.
A screenshot will not be made if the canvas size is too large, and an info dialog will be shown.

An additional quality setting that uses the minimum scale is added.
The resolutions are now calculated in terms of the maximum canvas dimensions.