Update base-zh-CN.yaml
Hi! I made corrections of some details in this version, and personally rephrased some sentences to make them more natural in Chinese. This is my first attempt to translate texts of a game, so some of my expressions may not be so precise, and my amendments may be somehow against translation rules. I am looking forward to your suggestions so that I can improve my translation skills.